MOve the movingAverage class to shared next to SimpleMovingAverage
